+++ frontend_server_common/lib/src/resident_runner.dart
@@ -18,6 +18,8 @@ import 'utilities.dart';
 
 final Uri platformDill =
     Uri.file(p.join(dartSdkPath, 'lib', '_internal', 'ddc_sdk.dill'));
+final Uri soundPlatformDill =
+  Uri.file(p.join(dartSdkPath, 'lib', '_internal', 'ddc_outline_sound.dill'));
 
 class ResidentWebRunner {
   ResidentWebRunner(
@@ -28,12 +30,14 @@ class ResidentWebRunner {
       this.fileSystemRoots,
       this.fileSystemScheme,
       this.outputPath,
+      this.soundNullSafety,
       bool verbose) {
     generator = ResidentCompiler(dartSdkPath,
         packagesPath: packagesPath,
-        platformDill: '$platformDill',
+        platformDill: soundNullSafety? '$soundPlatformDill': '$platformDill',
         fileSystemRoots: fileSystemRoots,
         fileSystemScheme: fileSystemScheme,
+        soundNullSafety: soundNullSafety,
         verbose: verbose);
     expressionCompiler = TestExpressionCompiler(generator);
   }
@@ -45,6 +49,7 @@ class ResidentWebRunner {
   final String outputPath;
   final List<String> fileSystemRoots;
   final String fileSystemScheme;
+  final bool soundNullSafety;
 
   ResidentCompiler generator;
   ExpressionCompiler expressionCompiler;
@@ -66,6 +71,7 @@ class ResidentWebRunner {
       packagesPath: packagesPath,
       root: root,
       urlTunneller: urlTunneller,
+      soundNullSafety: soundNullSafety,
     );
     uri = await devFS.create();