+++ dwds/CHANGELOG.md
@@ -10,9 +10,14 @@
 - Add `DwdsEvent`s around stepping and evaluation.
 - Send an event to the Dart Debug Extension that contains VM service protocol URI. 
 - Depend on `package:vm_service` version `6.1.0+1`.
+- Support `VMService.evaluate` using expression compiler.
 
 **Breaking changes:**
 - `LoadStrategy`s now require a `moduleInfoForEntrypoint`.
+  would cause a crash.
+- Change `ExpressionCompiler.initialize` method to include module format.
+- Add `LoadStrategy.moduleFormat` to be used for communicating current
+  module format to the expression compiler.
 
 ## 8.0.3